fix(review): rename tools_execution_helpers.py to _execution_helpers.py; add negative hasattr guards; update module count to 15
